pyproject.toml
[project]
name = "py-arch-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "django>=5.1.4,<5.2",
    "django-cors-headers>=4.6.0,<5.0.0",
    "django-ninja>=1.3.0,<2.0.0",
    "django-ninja-extra>=0.22.2,<0.23.0",
    "uvicorn[standard]",
    "sqlalchemy",
    "asyncpg",
    "python-dotenv",
    "email-validator",
    "pydantic",
    "pydantic-settings",
    "pytest-bdd-ng>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "passlib[bcrypt]>=1.7.4",
    "loguru>=0.7.3",
    "pyrefly>=0.35.0",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "pyright",
    "pytest",
    "pytest-xdist",
    "pytest-asyncio",
    "pytest-django",
    "httpx",
    "ruff",
    "pytest-bdd-ng>=2.3.1",
    "pytest-postgresql>=7.0.2",
    "psycopg-binary>=3.2.9",
]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "F", # pyflakes
    "B", # flake8-bugbear
    "I", # isort import sorting
    "W", # pycodestyle warnings
]
# If warning hints are annoying while coding, check their type and disable them
# They'll still be fixed when you save or run 'ruff check --fix .'
ignore = [
    "I001", # unsorted-imports 
    "E201", # ignore whitespace after '(' warning
    "E202", # ignore whitespace before ')' warning
    "E241", # allow multiple spaces for alignment
    "E272", # allow multiple spaces before keyword
    "E231", # missing whitespace after comma
    "W291", # trailing whitespace
    "W293", # line contains whitespace
    "E501", # ignore line too long warning
    "D107", # ignore missing docstring in __init__
    "C901", # too complex
    "B904", # Within an `except` clause, raise exceptions with `raise ... from err` or `raise ... from None` to distinguish them from errors in exception handling
    "B008", # Do not perform function call in argument defaults (needed for FastAPI Depends)
]
unfixable = ["F401", "F841"] # unused-import, unused-variable 

[tool.ruff.lint.isort]
order-by-type = true
force-sort-within-sections = true
force-single-line = false
lines-after-imports = 2
combine-as-imports = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false # preserve trailing commas

[tool.pytest.ini_options]
python_files = ["test.py", "test_*.py", "*_test.py"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
DJANGO_SETTINGS_MODULE = "src.platform.settings"
markers = [
    "asyncio: mark test as async",
    "bdd: BDD scenario test",
    "api: API test",
    "entity: Entity test",
    "integration: Integration test",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic._internal._config",
    "ignore::pytest.PytestReturnNotNoneWarning",
    "ignore::django.utils.deprecation.RemovedInDjango60Warning",
    "ignore::DeprecationWarning:gherkin.gherkin_line",
]
testpaths = ["test"]
# pytest-bdd configuration
bdd_features_base_dir = "test/features/"
addopts = "-v --tb=short --strict-markers"

[tool.pyright]

# some checking need to 
typeCheckingMode = "off"
reportMissingModuleSource = false
reportMissingImports = false
reportInvalidTypeForm = false

[tool.pyrefly]
search-path = [".", "test", ".venv/lib/python3.13/site-packages"]
site-package-path = [".venv/lib/python3.13/site-packages"]
project-includes = ["src", "test"]

[tool.pyrefly.errors]
missing-attribute = false
missing-argument = false
unexpected-keyword = false


[tool.uv]
package = false
